refactor(events): move out of src/constants
